enhancement: moved all hooks to its own folders
