+++ src/hooks/use-app-selector/index.ts
@@ -1,0 +1,1 @@
+export { default } from './useAppSelector';
