+++ src/hooks/use-get-course/index.ts
@@ -1,0 +1,1 @@
+export { default } from './useGetCourse';
