+++ src/hooks/use-get-courses/index.ts
@@ -1,0 +1,1 @@
+export { default } from './useGetCourses';
