+++ src/hooks/use-load-token/index.ts
@@ -1,0 +1,1 @@
+export { default } from './useLoadToken';
